Used DUs for `Dec`, `Bin` and `Hex`. Close #49.
`Dec`, `Bin` and `Hex` were represented as type of `int` or `string`.
However, in this way, incorrect values can get into processing.
In order to avoid it, there should be DUs for those types of numbers.

Now `Dec`, `Bin` and `Hex` are represented with single case DUs.

The signature and tests are writtes,

`isValid` functions and `tryToxxx` functions are removed
because of less use.