Dealt with the error in `Int32` and `Dec` modules. Close #68.
A test for `Int32.validate` and
another one for `Dec.validate` wouldn't pass.
That was because of the defferent .NET version.
An error message returned from `Int32.Parse` method was changed
starting from .NET 7,
and my `Int32.validate` depended on the message,
that was why the functions didn't pass.
Now they pass the test.